requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
ultralytics>=8.0.20
opencv-python>=4.7.0.72
matplotlib>=3.7.1
numpy>=1.25.0
wandb>=0.15.13
PyYAML>=6.0
scipy>=1.11.0
tdqm
onnx 
onnxslim 
onnxruntime
tensorrt